Document http interface's need for urlencode(dict)
This is to accomondate existing customers who accidentally depend on such behavior.
It also documents the future direction of using an already-encoded string,
so that such expectation is clear in this first release of this http interface.
In the future, when we switch existing dict behavior to string behavior,
it won't be considered as a breaking change.