Add close method to app
Fixes Issue #370 Session/socket never closed, warnings shown

The msal application does not close its resources like eg: self.http_client.
As a result a 'ResourceWarning unclosed <ssl.SSLSocket' is shown when the
app goes out of scope (when warnings are enabled)

This fix adds and explicit close method so developers can cleanly close
the app if they so desire.
In __init__ we trap and close the exception explicitely as users cannot
call close when __init__ fails.
Finally we don't close the http_client if it's externally managed.